Devide MyTest into 4 separate classes Minor updated to follow AAA pattern Added more than 5 datasets to ParametrizedTest
